[[FIX]] Track FutureReservedWords as idnts in ES5+
ES5 allowed a number identifiers which were previously classified as
FutureReserve to be used as binding identifiers. Ensure that when input
code defines such identifiers, usage counting operates as expected.

Although this patch is intended to correct a regression introduced by
commit 173f4634, it exposes a previously-existing bug. Prior to this
patch, JSHint did not flag FutureReservedWords when used as
IdentifierReferences. Correct this behavior since deferring a correction
to a dedicated commit would require additional and ultimately useless
logic.