[[FIX]] Interpret "object rest" ident as a binding
Previously, JSHint interpreted token sequences following the "object
rest" operator as expressions. As noted in the in-line comment, this was
done to improve the chance of recovery in response to a common
programming error. Unfortunately, it also caused valid syntax to be
misinterpreted--the expected identifier was treated as a
IdentifierReference instead of a BindingIdentifier.

Update the parsing strategy to use the expression parsing fallback only
when the input is known to be invalid.